User management functionality added

+++ Qtec.AccountManagement/Src/Qtec.AccountManagement.Web/Pages/Roles/RoleList.cshtml.cs
@@ -1,3 +1,5 @@
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Qtec.AccountManagement.Application.Services;
@@ -7,11 +9,14 @@ namespace Qtec.AccountManagement.Web.Pages.Roles
 {
     public class RoleListModel : PageModel
     {
+        private readonly IAuthorizationService _authorizationService;
         private readonly RoleManagementService _roleManagementService;
         private readonly ILogger<RoleListModel> _logger;
-        public RoleListModel(RoleManagementService roleManagementService, ILogger<RoleListModel> logger)
+        public RoleListModel(RoleManagementService roleManagementService, 
+            ILogger<RoleListModel> logger, IAuthorizationService authorizationService)
         {
             _roleManagementService = roleManagementService;
+            _authorizationService = authorizationService;
             _logger = logger;
         }
 
@@ -32,6 +37,12 @@ namespace Qtec.AccountManagement.Web.Pages.Roles
 
         public async Task<IActionResult> OnPostUpdateRoleAsync()
         {
+            var result = await _authorizationService.AuthorizeAsync(User, "Admin");
+
+            if (!result.Succeeded)
+            {
+                return RedirectToPage("/AccessDenied");
+            }
             var success = await _roleManagementService.UpdateRoleAsync(Id, Name);
 
             if (success)
@@ -44,6 +55,12 @@ namespace Qtec.AccountManagement.Web.Pages.Roles
 
         public async Task<IActionResult> OnPostDeleteRoleAsync(Guid Id)
         {
+            var result = await _authorizationService.AuthorizeAsync(User, "Admin");
+
+            if (!result.Succeeded)
+            {
+                return RedirectToPage("/AccessDenied");
+            }
             var success = await _roleManagementService.DeleteRoleAsync(Id);
             if (success)
             {
